+++ examples/pydantic_ai_examples/rag_lancedb.py
@@ -5,9 +5,6 @@ This example demonstrates a simpler RAG setup using:
 - chonkie for chunking documents
 - LanceDB as an embeddable vector database (no external service needed)
 
-This is a great starting point for RAG applications — everything runs locally
-without Docker or external databases.
-
 Build the search DB with:
 
     uv run -m pydantic_ai_examples.rag_lancedb build
@@ -17,12 +14,11 @@ Ask the agent a question with:
     uv run -m pydantic_ai_examples.rag_lancedb search "How do I configure logfire to work with FastAPI?"
 """
 
+# pyright: reportUnknownMemberType=false
 from __future__ import annotations as _annotations
 
 import asyncio
-import re
 import sys
-import unicodedata
 from dataclasses import dataclass
 from pathlib import Path
 
@@ -31,18 +27,17 @@ import lancedb
 import logfire
 import pyarrow as pa
 from chonkie import RecursiveChunker
-from pydantic import TypeAdapter
 
 from pydantic_ai import Agent, Embedder, RunContext
 
-# 'if-token-present' means nothing will be sent (and the example will work) if you don't have logfire configured
+from ._rag_common import DOCS_JSON, sections_ta
+
 logfire.configure(send_to_logfire='if-token-present')
 logfire.instrument_pydantic_ai()
 
-# Create an embedder using pydantic-ai's unified interface
 embedder = Embedder('openai:text-embedding-3-small')
+chunker = RecursiveChunker(chunk_size=512)
 
-# LanceDB stores data locally — no external service required
 DB_PATH = Path('.lancedb')
 TABLE_NAME = 'doc_sections'
 
@@ -69,7 +64,6 @@ async def retrieve(context: RunContext[Deps], search_query: str) -> str:
     ):
         result = await context.deps.embedder.embed_query(search_query)
 
-    # Search using LanceDB's vector search
     results = context.deps.table.search(result.embeddings[0]).limit(8).to_pandas()
 
     return '\n\n'.join(
@@ -92,19 +86,8 @@ async def run_agent(question: str):
 
 #######################################################
 # The rest of this file is dedicated to preparing the #
-# search database, and some utilities.                #
+# search database.                                    #
 #######################################################
-
-# JSON document from
-# https://gist.github.com/samuelcolvin/4b5bb9bb163b1122ff17e29e48c10992
-DOCS_JSON = (
-    'https://gist.githubusercontent.com/'
-    'samuelcolvin/4b5bb9bb163b1122ff17e29e48c10992/raw/'
-    '80c5925c42f1442c24963aaf5eb1a324d47afe95/logfire_docs.json'
-)
-
-# Create a chunker for splitting large content
-chunker = RecursiveChunker(chunk_size=512)
 
 
 async def build_search_db():
@@ -116,49 +99,31 @@ async def build_search_db():
     sections = sections_ta.validate_json(response.content)
 
     print(f'Processing {len(sections)} sections...')
-
-    # Prepare data for embedding
-    records = []
+    records: list[dict[str, str]] = []
     for section in sections:
-        url = section.url()
         content = section.embedding_content()
+        for chunk in chunker.chunk(content):
+            records.append(
+                {
+                    'url': section.url(),
+                    'title': section.title,
+                    'content': chunk.text,
+                }
+            )
 
-        # Chunk the content if it's large
-        chunks = chunker.chunk(content)
-        # Chunk the content if it's large
-        # For this example, we embed the first chunk if content was split
-        # In production, you might want to store multiple chunks per section
-        chunks = chunker.chunk(content)
-        chunk_text = chunks[0].text if chunks else content
-
-        records.append(
-            {
-                'url': url,
-                'title': section.title,
-                'content': section.content,
-                'chunk_text': chunk_text,
-            }
-        )
-
-    print('Generating embeddings...')
-    # Embed all chunks in batches
-    chunk_texts = [r['chunk_text'] for r in records]
-
-    # Process in batches of 100 to avoid rate limits
-    embeddings = []
+    print(f'Generating embeddings for {len(records)} chunks...')
+    chunk_texts = [r['content'] for r in records]
+    all_embeddings: list[list[float]] = []
     batch_size = 100
     for i in range(0, len(chunk_texts), batch_size):
         batch = chunk_texts[i : i + batch_size]
         with logfire.span(f'embed batch {i // batch_size + 1}'):
             result = await embedder.embed_documents(batch)
-            embeddings.extend(result.embeddings)
+            all_embeddings.extend(list(e) for e in result.embeddings)
         print(f'  Embedded {min(i + batch_size, len(chunk_texts))}/{len(chunk_texts)}')
 
-    # Create LanceDB table
     print('Creating LanceDB table...')
     db = lancedb.connect(DB_PATH)
-
-    # Define schema with vector dimension matching text-embedding-3-small
     schema = pa.schema(
         [
             pa.field('url', pa.string()),
@@ -167,57 +132,15 @@ async def build_search_db():
             pa.field('vector', pa.list_(pa.float32(), 1536)),
         ]
     )
-
-    # Prepare data with embeddings
     data = [
-        {
-            'url': r['url'],
-            'title': r['title'],
-            'content': r['content'],
-            'vector': list(emb),
-        }
-        for r, emb in zip(records, embeddings, strict=False)
+        {'url': r['url'], 'title': r['title'], 'content': r['content'], 'vector': emb}
+        for r, emb in zip(records, all_embeddings, strict=False)
     ]
 
-    # Create or overwrite table
     if TABLE_NAME in db.table_names():
         db.drop_table(TABLE_NAME)
-
     db.create_table(TABLE_NAME, data, schema=schema)
-    print(f'Created table with {len(data)} records at {DB_PATH}')
-
-
-@dataclass
-class DocsSection:
-    id: int
-    parent: int | None
-    path: str
-    level: int
-    title: str
-    content: str
-
-    def url(self) -> str:
-        url_path = re.sub(r'\.md$', '', self.path)
-        return (
-            f'https://logfire.pydantic.dev/docs/{url_path}/#{slugify(self.title, "-")}'
-        )
-
-    def embedding_content(self) -> str:
-        return '\n\n'.join((f'path: {self.path}', f'title: {self.title}', self.content))
-
-
-sections_ta = TypeAdapter(list[DocsSection])
-
-
-def slugify(value: str, separator: str, unicode: bool = False) -> str:
-    """Slugify a string, to make it URL friendly."""
-    # Taken unchanged from https://github.com/Python-Markdown/markdown/blob/3.7/markdown/extensions/toc.py#L38
-    if not unicode:
-        # Replace Extended Latin characters with ASCII, i.e. `žlutý` => `zluty`
-        value = unicodedata.normalize('NFKD', value)
-        value = value.encode('ascii', 'ignore').decode('ascii')
-    value = re.sub(r'[^\w\s-]', '', value).strip().lower()
-    return re.sub(rf'[{separator}\s]+', separator, value)
+    print(f'Created table with {len(data)} chunks at {DB_PATH}')
 
 
 if __name__ == '__main__':
